bug fixes
holy shit it actually works. i made a tx from scratch with a normal
bitcoin address AND a multisig address as inputs, and an address and
OP_RETURN as outputs. then i just did the sign() function for each of
the relevant private keys, and it parsed correctly and signed the right
things and constructed the final tx correctly.  sweet!